Change TargetPlatformMinVersion to minimal possible
Hello. Thanks for great library!
In my application I use handwritten code with same functional for cases, when user want to attach visual tree to bug report, so it should works in runtime, not only for e2e-tests.
My application works on 15053+, but I don't see any reason to not target 10240.